+++ src/app/(bsystems)/b-systems/(app)/crm/lead/[leadId]/page.tsx
@@ -321,6 +321,7 @@ export default async function BsLeadDetailPage({
               meetings={lead.meetings}
               proposals={lead.proposals}
               lostInfo={lead.lostInfo}
+              postponeInfos={lead.postponeInfos}
               won={lead.wonInfo}
             />
           </div>